requirements.txt
langchain-google-genai
langgraph
pydantic
typing-extensions
pandas
kagglehub
dotenv
sentence-transformers
faiss-cpu
numpy
tiktoken